refactor(app): move the state machine into ainb-app
AppState, its sections, the event reducer, the keymap, the session loader,
snapshot and event bus move to ainb-app/src/app. ainb-core/src/app keeps
UiState, the attach handler, the Screen trait and registry, and re-exports
the rest.

Types in ainb-app cannot name types in ainb-core, and inherent impls must
live with their type, so the move needs these seams:

- Key handlers take a crate-owned `Chord`. It gains a structured view
  (`Key`, `Mods`, `Chord::new`, `code`, `modifiers`), and the one crossterm
  conversion, `terminal_keys::chord_from_key_event`, stays in core. It
  returns None for keys the keymap cannot spell. The old conversion
  panicked on those, for example Ctrl+Space (`KeyCode::Null`).
- `KeyHost` stands in for `&mut UiState` in key dispatch; UiState
  implements it.
- `SessionsPaneHitTest` and `PluginViewports` replace the two other UiState
  reads in AppState.
- The mouse reducers read the rects UiState records, so they move out of
  events.rs into core app/mouse.rs.
- The component halves P1b deferred because they reach AppState move too:
  session_recovery state, session_tabs, and the configure and pick_repo key
  handlers.

build.rs follows state.rs to its new path for the render-thread `.await`
lint. A missing scan file now fails the build instead of passing
vacuously.
