+++ ainb-tui/crates/ainb-app/src/app/event_bus.rs
