+++ ainb-tui/crates/ainb-app/src/app/sections.rs
