+++ ainb-tui/crates/ainb-app/src/app/session_loader.rs
