+++ ainb-tui/crates/ainb-app/src/app/snapshot.rs
