+++ ainb-tui/crates/ainb-core/src/app/mod.rs
@@ -1,23 +1,17 @@
-// ABOUTME: Main application structure and state management for the TUI
+// ABOUTME: Terminal half of the application: UiState (scroll, geometry, mouse),
+// the tty attach handler and the screen registry. The state machine itself lives
+// in `ainb_app::app` and is re-exported here, so `crate::app::state` and friends
+// keep resolving.
+
+pub use ainb_app::app::*;
 
 pub mod attach_handler;
-pub mod event_bus;
-pub mod events;
-pub mod keymap;
-pub mod keymap_defaults;
-pub mod keymap_toml;
+pub mod mouse;
 pub mod registry;
 pub mod screens;
-pub mod sections;
-pub mod session_loader;
-pub mod snapshot;
-pub mod state;
+pub mod terminal_keys;
 pub mod ui_state;
-pub mod versioned;
 
 pub use attach_handler::AttachHandler;
-pub use events::EventHandler;
 pub use registry::ScreenRegistry;
 pub use screens::{Screen, ScreenId};
-pub use session_loader::SessionLoader;
-pub use state::{App, AppState};
