+++ ainb-tui/crates/ainb-core/src/app/screens/builtin.rs
@@ -1,5 +1,8 @@
 // ABOUTME: Built-in Screen impls — thin shims around existing component renderers; no logic moves in Phase 2a
 
+// Screen-to-plugin routing lives in `ainb_app::app::screens::builtin`.
+pub use ainb_app::app::screens::builtin::*;
+
 use crate::app::ui_state::UiState;
 use ratatui::{Frame, layout::Rect};
 
@@ -55,25 +58,6 @@ impl PluginScreen {
     pub fn new(screen_id: &'static str) -> Self {
         Self { screen_id }
     }
-}
-
-/// Static screen → plugin routing table. The `state.rs` render tick
-/// already maps the same way; both call sites read this so the
-/// authoritative list lives in one place.
-///
-/// Keep this list in sync with `tick_plugin_renders` in `app/state.rs`.
-pub const PLUGIN_SCREENS: &[(&str, &str)] = &[
-    (ids::ANALYTICS, "burndown"),
-    (ids::WITR, "witr"),
-    (ids::LEARNINGS, "learnings"),
-    (ids::ABTOP, "abtop"),
-    (ids::HANGAR, "hangar-tui"),
-];
-
-/// Resolve the plugin id that owns `screen_id`, if any.
-#[must_use]
-pub fn plugin_id_for_screen(screen_id: &str) -> Option<&'static str> {
-    PLUGIN_SCREENS.iter().find_map(|(s, p)| (*s == screen_id).then_some(*p))
 }
 
 /// Convert a `crossterm::event::KeyEvent` into the portable wire
@@ -134,26 +118,6 @@ pub fn crossterm_to_protocol_key(
     Some(ProtocolEvent { code, mods, kind })
 }
 
-/// `true` when the plugin owning `current_screen` reported (on its last frame)
-/// that its focused surface is capturing free text — a title/filter/compose/
-/// search/API-key input where every printable key is typed content.
-///
-/// Read on the host key-dispatch path so `?`/`H`/`W` reach the plugin's input
-/// verbatim instead of toggling help / wiring the statusline (8hx). The flag is
-/// refreshed every tick by `AppState::tick_plugin_renders` from the plugin's
-/// `RenderResult.captures_text`. Non-plugin screens (and screens whose plugin
-/// has never painted) read `false`.
-#[must_use]
-pub fn focused_plugin_captures_text(state: &AppState) -> bool {
-    plugin_id_for_screen(&state.shell.current_screen).is_some()
-        && state
-            .plugins_host
-            .plugin_captures_text
-            .get(&state.shell.current_screen)
-            .copied()
-            .unwrap_or(false)
-}
-
 /// `true` if the host reserves this key — it MUST NOT be forwarded to
 /// the plugin and MUST fall through to the central dispatch.
 ///
@@ -197,21 +161,6 @@ pub fn is_host_reserved_key(
         CtKey::Char('?' | 'H') => !plugin_owns_help && !capturing,
         _ => false,
     }
-}
-
-/// Plugins that render their own `?` help overlay. On their screens the host
-/// never claims `?`/`H` (or the `W` statusline global), whatever the per-frame
-/// `captures_text` flag says. Every other plugin keeps the host help toggle.
-pub const PLUGINS_WITH_OWN_HELP: &[&str] = &["hangar-tui"];
-
-/// `true` when the focused screen belongs to a plugin in
-/// [`PLUGINS_WITH_OWN_HELP`]. The host's printable-key globals (`?`/`H` help,
-/// `W` statusline) are suppressed there; see [`is_host_reserved_key`] for why
-/// the per-frame `captures_text` flag alone is not a safe gate.
-#[must_use]
-pub fn plugin_owns_help_keys(state: &AppState) -> bool {
-    plugin_id_for_screen(&state.shell.current_screen)
-        .is_some_and(|id| PLUGINS_WITH_OWN_HELP.contains(&id))
 }
 
 /// Try to forward `key` to the plugin owning `current_screen`. Returns
@@ -368,7 +317,8 @@ pub fn forward_mouse_to_focused_plugin(
         .copied()
         .unwrap_or((0, 0));
     let area = ui
-        .plugin_render_areas
+        .plugin_viewports
+        .render_areas
         .get(&state.shell.current_screen)
         .copied()
         .unwrap_or((0, 0));
@@ -648,7 +598,8 @@ impl Screen for PluginScreen {
         // `App::tick_plugin_renders` can ask the plugin for a buffer that
         // actually fills this area. Without this the plugin renders into
         // its fallback (80×24) and everything outside that rect stays blank.
-        ui.plugin_render_areas
+        ui.plugin_viewports
+            .render_areas
             .insert(self.screen_id.to_string(), (area.width, area.height));
         // Stash the origin too, so the mouse forwarder can translate an
         // absolute terminal click into this plugin's viewport space.
