+++ ainb-tui/crates/ainb-core/src/components/new_session/pick_repo.rs
@@ -10,7 +10,6 @@
 
 pub use ainb_app::components::new_session::pick_repo::*;
 
-use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
 use ratatui::{
     Frame,
     layout::{Alignment, Constraint, Direction, Layout, Rect},
@@ -35,16 +34,6 @@ const SOFT_WHITE: Color = Color::Rgb(220, 220, 230);
 const MUTED_GRAY: Color = Color::Rgb(120, 120, 140);
 const DARK_BG: Color = Color::Rgb(25, 25, 35);
 const LIST_HIGHLIGHT_BG: Color = Color::Rgb(40, 40, 60);
-
-/// Result of toggling a favorite — drives the `^F` notification.
-enum FavoriteToggle {
-    /// Newly favorited; carries the stored remote source for display.
-    Added(String),
-    /// Un-favorited; carries the row label for display.
-    Removed(String),
-    /// Refused because the row has no remote repository indicator.
-    Refused(String),
-}
 
 /// A dimmed locator shown after the row name so identically-named repos are
 /// distinguishable. Derived from the row's `RepoSource`: a `~`-abbreviated path
@@ -384,288 +373,10 @@ fn render_git_auth_modal(f: &mut Frame, state: &PickRepoState, area: Rect) {
     f.render_widget(Paragraph::new(cta).alignment(Alignment::Center), parts[1]);
 }
 
-/// Handle a single key event. Mutates state in place; returns the outcome
-/// the caller should act on (advance, return home, start clone, or stay).
-///
-/// Persistence is the dispatcher's job for navigation keys (finding #3 +
-/// #11) — arrow/Esc/Enter only mutate the in-memory `state.defaults`
-/// snapshot, and the dispatcher writes to
-/// `~/.agents-in-a-box/session-defaults.yaml` when the screen exits
-/// (AdvanceTo / StartClone / BackToHome). Two exceptions:
-///   1. `^R` (reset) — a deliberate user-issued clear; we persist
-///      synchronously here so the next Esc doesn't immediately re-record
-///      a sticky-cursor highlight that the user just told us to wipe.
-///   2. `^F` (favorite toggle) — already writes `favorites.yaml`
-///      synchronously inside `toggle_favorite`.
-pub fn handle_key(state: &mut PickRepoState, key: KeyEvent) -> PickRepoOutcome {
-    // When an auth check is in flight or failed, intercept keys before
-    // normal picker handling. Checking → only Esc; NotAuthenticated →
-    // Enter retries, s skips, Esc clears.
-    if let Some(ref auth) = state.git_auth_status {
-        match auth {
-            GitAuthStatus::Checking => {
-                if matches!(key.code, KeyCode::Esc) {
-                    state.git_auth_status = None;
-                    state.pending_clone_source = None;
-                    state.git_auth_error = None;
-                }
-                return PickRepoOutcome::Stay;
-            }
-            GitAuthStatus::NotAuthenticated => {
-                match key.code {
-                    KeyCode::Enter => {
-                        // Re-probe: drop the stale error so the modal shows the
-                        // spinner, not last attempt's failure, while it re-runs.
-                        state.git_auth_error = None;
-                        state.git_auth_status = Some(GitAuthStatus::Checking);
-                        return PickRepoOutcome::Stay; // dispatcher sees Checking → re-runs check
-                    }
-                    KeyCode::Char('s' | 'S') => {
-                        let source = state.pending_clone_source.take();
-                        state.git_auth_status = None;
-                        state.git_auth_error = None;
-                        if let Some(src) = source {
-                            return PickRepoOutcome::StartClone(src);
-                        }
-                        return PickRepoOutcome::Stay;
-                    }
-                    KeyCode::Esc => {
-                        state.git_auth_status = None;
-                        state.pending_clone_source = None;
-                        state.git_auth_error = None;
-                        return PickRepoOutcome::Stay;
-                    }
-                    _ => return PickRepoOutcome::Stay,
-                }
-            }
-            GitAuthStatus::Authenticated => {
-                // Auto-advance handled by dispatcher; shouldn't linger here
-                state.git_auth_status = None;
-            }
-        }
-    }
-
-    let defaults_path = SessionDefaults::default_path();
-    // Ctrl-modified keys take precedence over plain chars so `^R` / `^F`
-    // never get swallowed by the filter-input branch below.
-    if key.modifiers.contains(KeyModifiers::CONTROL) {
-        match key.code {
-            KeyCode::Char('r' | 'R') => {
-                tracing::debug!("pick_repo: ^R reset");
-                state.filter.clear();
-                state.defaults.reset_last_repo();
-                state.refilter();
-                state.selected = 0;
-                // Persist the cleared state immediately — the next Esc
-                // would otherwise re-stamp `last_repo` with the row 0 id
-                // (sticky-cursor UX) and silently undo the ^R.
-                if let Err(err) = state.defaults.save_to(&defaults_path) {
-                    tracing::warn!(error = %err, "pick_repo: ^R persist failed");
-                }
-                return PickRepoOutcome::Stay;
-            }
-            KeyCode::Char('f' | 'F') => {
-                tracing::debug!("pick_repo: ^F toggle favorite on highlighted row");
-                if let Some(row) = state.highlighted().cloned() {
-                    return match toggle_favorite(state, &row) {
-                        FavoriteToggle::Refused(reason) => PickRepoOutcome::Notice {
-                            message: format!("★ Can't favorite: {reason}"),
-                            is_error: true,
-                        },
-                        FavoriteToggle::Added(display) => {
-                            let local_repos = collect_local_repo_paths(state);
-                            state.rebuild_rows(&local_repos);
-                            PickRepoOutcome::Notice {
-                                message: format!("⭐ Added '{display}' to favorites"),
-                                is_error: false,
-                            }
-                        }
-                        FavoriteToggle::Removed(display) => {
-                            let local_repos = collect_local_repo_paths(state);
-                            state.rebuild_rows(&local_repos);
-                            PickRepoOutcome::Notice {
-                                message: format!("★ Removed '{display}' from favorites"),
-                                is_error: false,
-                            }
-                        }
-                    };
-                }
-                return PickRepoOutcome::Stay;
-            }
-            KeyCode::Char('v' | 'V') => {
-                // Ctrl+V: ask the caller to read the OS clipboard and append
-                // it (Cmd+V / bracketed paste isn't delivered to this field
-                // in some terminals — tmux, mouse-capture).
-                tracing::debug!("pick_repo: ^V clipboard paste");
-                return PickRepoOutcome::PasteFromClipboard;
-            }
-            _ => {}
-        }
-    }
-
-    match key.code {
-        KeyCode::Esc => {
-            // Esc with text typed → clear filter first; on second press
-            // (empty filter) → return to home. Sticky-cursor highlight is
-            // stored in-memory only; dispatcher persists on exit.
-            if !state.filter.is_empty() {
-                if let Some(row) = state.highlighted() {
-                    state.defaults.last_repo = Some(row.id.clone());
-                }
-                state.filter.clear();
-                state.refilter();
-                return PickRepoOutcome::Stay;
-            }
-            if let Some(row) = state.highlighted() {
-                state.defaults.last_repo = Some(row.id.clone());
-            }
-            PickRepoOutcome::BackToHome
-        }
-        KeyCode::Up => {
-            if !state.filtered_indices.is_empty() {
-                state.selected = if state.selected == 0 {
-                    state.filtered_indices.len() - 1
-                } else {
-                    state.selected - 1
-                };
-                if let Some(row) = state.highlighted() {
-                    state.defaults.last_repo = Some(row.id.clone());
-                }
-            }
-            PickRepoOutcome::Stay
-        }
-        KeyCode::Down => {
-            if !state.filtered_indices.is_empty() {
-                state.selected = (state.selected + 1) % state.filtered_indices.len();
-                if let Some(row) = state.highlighted() {
-                    state.defaults.last_repo = Some(row.id.clone());
-                }
-            }
-            PickRepoOutcome::Stay
-        }
-        KeyCode::Enter => {
-            // Prefer a row hit; fall back to smart-parse on the filter text.
-            if let Some(row) = state.highlighted().cloned() {
-                tracing::debug!("pick_repo: Enter on row {}", row.id);
-                state.defaults.last_repo = Some(row.id.clone());
-                return resolve_outcome(row.source);
-            }
-            // No matches — smart-parse the filter as raw input.
-            if state.filter.is_empty() {
-                return PickRepoOutcome::Stay;
-            }
-            let parsed = parse_with(&state.filter, &RealFs);
-            tracing::debug!("pick_repo: smart-parse {:?} -> {parsed:?}", state.filter);
-            state.defaults.last_repo = Some(state.filter.clone());
-            resolve_outcome(parsed)
-        }
-        KeyCode::Backspace => {
-            state.filter.pop();
-            state.refilter();
-            PickRepoOutcome::Stay
-        }
-        KeyCode::Char(c) if !key.modifiers.contains(KeyModifiers::CONTROL) => {
-            state.filter.push(c);
-            state.refilter();
-            PickRepoOutcome::Stay
-        }
-        _ => PickRepoOutcome::Stay,
-    }
-}
-
-/// Dispatch a resolved `RepoSource` into a `PickRepoOutcome` per spec table:
-/// - `LocalPath` / pre-cloned `GithubShorthand` (local hit) → `AdvanceTo`
-/// - `HttpsUrl` / `SshUrl` / `GithubShorthand` (remote) → `StartClone`
-/// - `SshSession` → `AdvanceTo` (no clone needed)
-/// - `Filter` → `Stay` (just an unparseable string)
-fn resolve_outcome(source: RepoSource) -> PickRepoOutcome {
-    match source {
-        RepoSource::LocalPath(_) | RepoSource::SshSession(_) => PickRepoOutcome::AdvanceTo(source),
-        RepoSource::HttpsUrl(_) | RepoSource::SshUrl(_) | RepoSource::GithubShorthand { .. } => {
-            PickRepoOutcome::StartClone(source)
-        }
-        RepoSource::Filter(_) => PickRepoOutcome::Stay,
-    }
-}
-
-/// Toggle the favorite status of a row. Updates both the in-memory store
-/// (mutating `state.favorites`) AND the on-disk YAML so the change survives
-/// across TUI restarts.
-///
-/// A favorite ALWAYS records a remote indicator. Remote rows store directly;
-/// a `LocalPath` row is resolved to its `origin` remote via
-/// [`favorite_from_local_repo`] (refused if it has none). `SshSession`
-/// (interactive, not a repo) and `Filter` (unparseable text) rows are refused
-/// outright — never persisted.
-fn toggle_favorite(state: &mut PickRepoState, row: &PickRepoRow) -> FavoriteToggle {
-    if state.favorites.has_alias(&row.id) {
-        state.favorites.remove(&row.id);
-        persist_favorites(state);
-        return FavoriteToggle::Removed(row.label.clone());
-    }
-
-    let fav = match &row.source {
-        RepoSource::HttpsUrl(u) => Favorite::new(row.id.clone(), u.clone(), SourceType::HttpsUrl),
-        RepoSource::SshUrl(u) => Favorite::new(row.id.clone(), u.clone(), SourceType::SshUrl),
-        RepoSource::GithubShorthand { owner, repo } => Favorite::new(
-            row.id.clone(),
-            format!("{owner}/{repo}"),
-            SourceType::GithubShorthand,
-        ),
-        RepoSource::LocalPath(p) => {
-            match crate::config::favorite_from_local_repo(row.id.clone(), p) {
-                Ok(fav) => fav,
-                Err(e) => {
-                    tracing::warn!(
-                        alias = %row.id,
-                        path = %p.display(),
-                        error = %e,
-                        "pick_repo: refusing to favorite local row — no remote indicator",
-                    );
-                    return FavoriteToggle::Refused(e.to_string());
-                }
-            }
-        }
-        RepoSource::SshSession(s) | RepoSource::Filter(s) => {
-            tracing::warn!(
-                alias = %row.id,
-                text = %s,
-                "pick_repo: refusing to favorite — not a remote repository indicator",
-            );
-            return FavoriteToggle::Refused("not a remote repository".to_string());
-        }
-    };
-
-    let display = fav.source.clone();
-    state.favorites.set(fav);
-    persist_favorites(state);
-    FavoriteToggle::Added(display)
-}
-
-/// Persist the favorites store to disk, logging (but not failing) on error.
-fn persist_favorites(state: &PickRepoState) {
-    if let Err(err) = state.favorites.save() {
-        tracing::warn!(error = %err, "pick_repo: failed to persist favorites");
-    }
-}
-
-/// Pull current local-only paths out of state so the row list can be
-/// rebuilt without losing the local-scan input.
-fn collect_local_repo_paths(state: &PickRepoState) -> Vec<PathBuf> {
-    state
-        .rows
-        .iter()
-        .filter(|r| r.kind == RowKind::Local)
-        .filter_map(|r| match &r.source {
-            RepoSource::LocalPath(p) => Some(p.clone()),
-            _ => None,
-        })
-        .collect()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::app::keymap::{Chord, Key, Mods};
     use crate::config::favorites_store::FavoritesStore;
     use crate::config::favorites_store::SourceType;
 
@@ -858,9 +569,9 @@ mod tests {
     #[test]
     fn ctrl_v_requests_clipboard_paste() {
         let mut s = mk_state_with_rows(one_row());
-        let ctrl_v = KeyEvent::new(KeyCode::Char('v'), KeyModifiers::CONTROL);
+        let ctrl_v = Chord::new(Key::Char('v'), Mods::CTRL);
         assert_eq!(
-            handle_key(&mut s, ctrl_v),
+            handle_key(&mut s, &ctrl_v),
             PickRepoOutcome::PasteFromClipboard
         );
         // The keystroke must NOT also type a literal 'v' into the filter.
@@ -870,8 +581,8 @@ mod tests {
     #[test]
     fn plain_v_still_types_into_filter() {
         let mut s = mk_state_with_rows(one_row());
-        let v = KeyEvent::new(KeyCode::Char('v'), KeyModifiers::NONE);
-        assert_eq!(handle_key(&mut s, v), PickRepoOutcome::Stay);
+        let v = Chord::new(Key::Char('v'), Mods::NONE);
+        assert_eq!(handle_key(&mut s, &v), PickRepoOutcome::Stay);
         assert_eq!(s.filter, "v");
     }
 
@@ -1023,7 +734,7 @@ mod tests {
         s.git_auth_status = Some(GitAuthStatus::NotAuthenticated);
         s.pending_clone_source = Some(github_source());
 
-        let out = handle_key(&mut s, KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE));
+        let out = handle_key(&mut s, &Chord::new(Key::Enter, Mods::NONE));
 
         // Enter flips back to Checking so the dispatcher re-runs `gh auth status`.
         assert_eq!(out, PickRepoOutcome::Stay);
@@ -1038,10 +749,7 @@ mod tests {
         s.git_auth_status = Some(GitAuthStatus::NotAuthenticated);
         s.pending_clone_source = Some(github_source());
 
-        let out = handle_key(
-            &mut s,
-            KeyEvent::new(KeyCode::Char('s'), KeyModifiers::NONE),
-        );
+        let out = handle_key(&mut s, &Chord::new(Key::Char('s'), Mods::NONE));
 
         // `s` bypasses the gate and proceeds to clone with the pending source.
         assert_eq!(out, PickRepoOutcome::StartClone(github_source()));
@@ -1055,7 +763,7 @@ mod tests {
         s.git_auth_status = Some(GitAuthStatus::NotAuthenticated);
         s.pending_clone_source = Some(github_source());
 
-        let out = handle_key(&mut s, KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE));
+        let out = handle_key(&mut s, &Chord::new(Key::Esc, Mods::NONE));
 
         // Esc dismisses the prompt and clears the pending source.
         assert_eq!(out, PickRepoOutcome::Stay);
@@ -1070,15 +778,12 @@ mod tests {
         s.pending_clone_source = Some(github_source());
 
         // A stray keypress while the check is in flight is ignored, state holds.
-        let out = handle_key(
-            &mut s,
-            KeyEvent::new(KeyCode::Char('x'), KeyModifiers::NONE),
-        );
+        let out = handle_key(&mut s, &Chord::new(Key::Char('x'), Mods::NONE));
         assert_eq!(out, PickRepoOutcome::Stay);
         assert_eq!(s.git_auth_status, Some(GitAuthStatus::Checking));
 
         // Esc aborts the in-flight check and drops the pending source.
-        let out = handle_key(&mut s, KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE));
+        let out = handle_key(&mut s, &Chord::new(Key::Esc, Mods::NONE));
         assert_eq!(out, PickRepoOutcome::Stay);
         assert_eq!(s.git_auth_status, None);
         assert_eq!(s.pending_clone_source, None);
