+++ ainb-tui/crates/ainb-core/tests/home_sidebar_mouse.rs
@@ -27,16 +27,22 @@ fn home_sidebar_mouse_click_selects_and_double_click_navigates() {
     // Sidebar rect starts at y=4, so first item row is y=7. With Sessions
     // (index 0) selected and thus 2 rows tall, y=10 lands on index 2 =
     // Config per SidebarItem::all().
-    let first =
-        EventHandler::handle_mouse_event(AppEvent::MouseClick { x: 3, y: 10 }, &mut state, &mut ui);
+    let first = ainb::app::mouse::handle_mouse_event(
+        AppEvent::MouseClick { x: 3, y: 10 },
+        &mut state,
+        &mut ui,
+    );
     assert!(first.is_none());
     assert_eq!(
         state.shell.home_screen_v2_state.sidebar.selected_item(),
         SidebarItem::Config
     );
 
-    let second =
-        EventHandler::handle_mouse_event(AppEvent::MouseClick { x: 3, y: 10 }, &mut state, &mut ui);
+    let second = ainb::app::mouse::handle_mouse_event(
+        AppEvent::MouseClick { x: 3, y: 10 },
+        &mut state,
+        &mut ui,
+    );
     assert!(matches!(second, Some(AppEvent::HomeScreenSidebarSelect)));
 }
 
@@ -54,7 +60,7 @@ fn home_sidebar_resize_release_persists_width_to_isolated_home() {
     state.shell.home_screen_v2_state.last_sidebar_rect =
         Some(ainb::geometry::Area::new(0, 4, 26, 30));
 
-    let down = EventHandler::handle_mouse_event(
+    let down = ainb::app::mouse::handle_mouse_event(
         AppEvent::MouseClick { x: 25, y: 10 },
         &mut state,
         &mut ui,
@@ -62,7 +68,7 @@ fn home_sidebar_resize_release_persists_width_to_isolated_home() {
     assert!(down.is_none());
     assert!(state.shell.home_screen_v2_state.sidebar_resize_active);
 
-    let drag = EventHandler::handle_mouse_event(
+    let drag = ainb::app::mouse::handle_mouse_event(
         AppEvent::MouseDragging { x: 29, y: 10 },
         &mut state,
         &mut ui,
@@ -70,7 +76,7 @@ fn home_sidebar_resize_release_persists_width_to_isolated_home() {
     assert!(drag.is_none());
     assert_eq!(state.shell.home_screen_v2_state.sidebar.preferred_width, 30);
 
-    let up = EventHandler::handle_mouse_event(
+    let up = ainb::app::mouse::handle_mouse_event(
         AppEvent::MouseDragEnd { x: 29, y: 10 },
         &mut state,
         &mut ui,
