+++ ainb-tui/crates/ainb-core/tests/test_manual_refresh.rs
@@ -28,7 +28,10 @@ async fn test_manual_refresh_key() {
 
     // Simulate pressing 'f' for refresh
     let refresh_key = KeyEvent::new(KeyCode::Char('f'), KeyModifiers::NONE);
-    if let Some(event) = EventHandler::handle_key_event(refresh_key, &mut app.state) {
+    if let Some(event) = EventHandler::handle_key_event(
+        ainb::app::terminal_keys::chord_from_key_event(&refresh_key).expect("mapped key"),
+        &mut app.state,
+    ) {
         EventHandler::process_event(event, &mut app.state);
     }
 
@@ -76,7 +79,10 @@ async fn test_refresh_from_session_list_view() {
 
     // Press 'f' to refresh
     let refresh_key = KeyEvent::new(KeyCode::Char('f'), KeyModifiers::NONE);
-    if let Some(event) = EventHandler::handle_key_event(refresh_key, &mut app.state) {
+    if let Some(event) = EventHandler::handle_key_event(
+        ainb::app::terminal_keys::chord_from_key_event(&refresh_key).expect("mapped key"),
+        &mut app.state,
+    ) {
         EventHandler::process_event(event, &mut app.state);
     }
 
@@ -128,7 +134,10 @@ async fn test_multiple_refreshes() {
     for i in 0..3 {
         // Press 'f' to refresh
         let refresh_key = KeyEvent::new(KeyCode::Char('f'), KeyModifiers::NONE);
-        if let Some(event) = EventHandler::handle_key_event(refresh_key, &mut app.state) {
+        if let Some(event) = EventHandler::handle_key_event(
+            ainb::app::terminal_keys::chord_from_key_event(&refresh_key).expect("mapped key"),
+            &mut app.state,
+        ) {
             EventHandler::process_event(event, &mut app.state);
         }
 
@@ -166,14 +175,20 @@ async fn test_refresh_doesnt_interfere_with_help() {
 
     // Show help first
     let help_key = KeyEvent::new(KeyCode::Char('?'), KeyModifiers::NONE);
-    if let Some(event) = EventHandler::handle_key_event(help_key, &mut app.state) {
+    if let Some(event) = EventHandler::handle_key_event(
+        ainb::app::terminal_keys::chord_from_key_event(&help_key).expect("mapped key"),
+        &mut app.state,
+    ) {
         EventHandler::process_event(event, &mut app.state);
     }
     assert!(app.state.shell.help_visible);
 
     // Try to refresh while help is visible - should not trigger refresh in help mode
     let refresh_key = KeyEvent::new(KeyCode::Char('f'), KeyModifiers::NONE);
-    let event_option = EventHandler::handle_key_event(refresh_key, &mut app.state);
+    let event_option = EventHandler::handle_key_event(
+        ainb::app::terminal_keys::chord_from_key_event(&refresh_key).expect("mapped key"),
+        &mut app.state,
+    );
 
     // In help mode, 'f' should not trigger refresh
     assert!(
@@ -183,14 +198,20 @@ async fn test_refresh_doesnt_interfere_with_help() {
 
     // Close help
     let esc_key = KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE);
-    if let Some(event) = EventHandler::handle_key_event(esc_key, &mut app.state) {
+    if let Some(event) = EventHandler::handle_key_event(
+        ainb::app::terminal_keys::chord_from_key_event(&esc_key).expect("mapped key"),
+        &mut app.state,
+    ) {
         EventHandler::process_event(event, &mut app.state);
     }
     assert!(!app.state.shell.help_visible);
 
     // Now refresh should work
     let refresh_key = KeyEvent::new(KeyCode::Char('f'), KeyModifiers::NONE);
-    if let Some(event) = EventHandler::handle_key_event(refresh_key, &mut app.state) {
+    if let Some(event) = EventHandler::handle_key_event(
+        ainb::app::terminal_keys::chord_from_key_event(&refresh_key).expect("mapped key"),
+        &mut app.state,
+    ) {
         EventHandler::process_event(event, &mut app.state);
     }
