+++ ainb-tui/crates/ainb-core/tests/test_session_creation_refresh.rs
@@ -34,7 +34,10 @@ async fn app_on_picker_from_session_list() -> App {
     app.state.shell.current_screen = screen_ids::SESSION_LIST.to_string();
 
     let key_event = KeyEvent::new(KeyCode::Char('n'), KeyModifiers::NONE);
-    if let Some(event) = EventHandler::handle_key_event(key_event, &mut app.state) {
+    if let Some(event) = EventHandler::handle_key_event(
+        ainb::app::terminal_keys::chord_from_key_event(&key_event).expect("mapped key"),
+        &mut app.state,
+    ) {
         EventHandler::process_event(event, &mut app.state);
     }
     // The picker opens synchronously and only spawns a background rescan; the
